Floats lead to jagged and low accuracy math functions after just 1000 seconds or so. So move back to Double. It's a shame to do that rather than ensure that time is modulused or another solution that doesn't double the size of all the time and signal values. Also, in LowPassFilter2 track the absolute time better.

+++ ProgressionPlayer/Sources/Tones/Arrow.swift
@@ -7,7 +7,7 @@
 
 import AVFAudio
 
-typealias CoreFloat = Float
+typealias CoreFloat = Double
 
 class Arrow11 {
   // these are arrows with which we can compose (arr/arrs run first, then this arrow)
